move clients around
