+++ config/graphqlClients.ts
@@ -1,12 +1,10 @@
-import {
-  NEXT_PUBLIC_PSYC_SUBGRAPH_URL,
-  SNAPSHOT_GRAPHQL_URL
-} from "@/constants/claims";
+import { env } from "@/config/env.mjs";
+import { SNAPSHOT_GRAPHQL_URL } from "@/constants/claims";
 import { GraphQLClient } from "graphql-request";
 
 export const snapshotGraphQLClient = new GraphQLClient(
   SNAPSHOT_GRAPHQL_URL as string
 );
 export const psycGraphQLClient = new GraphQLClient(
-  NEXT_PUBLIC_PSYC_SUBGRAPH_URL as string
+  env.NEXT_PUBLIC_SUBGRAPH_URL as string
 );
