+++ hooks/useBuyNft.ts
@@ -13,7 +13,7 @@ import { useResize } from "./useResize";
 import { psycSaleContractConfig } from "@/lib/sale-contract-config";
 import { toWei } from "@/utils/saleUtils";
 import { useCustomToasts } from "./useCustomToasts";
-import graphClient from "@/config/graphql";
+import { graphClient } from "@/config/apolloClients";
 
 type ArgsType =
   | [number, string[]]
@@ -128,6 +128,7 @@ const useBuyNft = (
       setIsModalOpen(false);
     } else if (isSuccess) {
       handleTransactionSuccess(width);
+
       graphClient.cache.evict({
         fieldName: "userCopyBalance",
         broadcast: false
